All UIElements now should have their parents from construction. Talents implementation have been started.

+++ Project 1/World/GameObjects/Unit/Resources/Energy.cs
@@ -55,7 +55,7 @@ namespace Project_1.GameObjects.Unit.Resources
             regenValue = 20.0f;
         }
 
-        public override void CastSpell(float aCost)
+        public override void CastSpell(double aCost)
         {
             AssertSimThread();
             base.CastSpell(aCost);
